Refactor moving Kotlin instantiator package
